Azure/blob-storage/functions/scanner/requirements.txt
azure-functions
azure-storage-blob
azure-storage-queue
grpcio>=1.60.0
protobuf>=4.25.0
